BSP 3.0
Changes in 3.0x:

Lee Killough 9/98:

Fixed many cases of slime lines by going back to using slightly slower BSP21x
splitting algorithm.

Removed approximate visplane overflow detection code, as being too inaccurate
to be useful. If visplane overflows need to be detected, then modifying the
Doom source code to display them exactly in real-time would be a better idea.
The visplane counting code in BSP22x and BSP23x was never 100% accurate, but
was about as close as I could make it based solely on empirical evidence.

NOTE: this change does not remove the -vp option, which attempts to reduce
the chances of visplane overflows by using a special node line heuristic, and
which is the more useful of the visplane-related options. This change only
removes the -vpwarn and -vpmark options, which attempted to detect where
visplane overflows occurred, but which were too crude. Exact detection of
visplane overflows is unnecessary to remove them, because techniques are
available which reduce the probability of visplane overflows without having
to know whether they actually occur.

+++ bsp.h
@@ -26,7 +26,7 @@
 static void ProgError( char *, ...);
 static void *GetMemory(size_t);
 static void *ResizeMemory(void *, size_t);
-static unsigned ComputeAngle(long,long);
+static unsigned ComputeAngle(int,int);
 
 #undef max
 #define max(a,b) (((a)>(b))?(a):(b))
